refactor: Move MenuService to /domain/menu/service (#23)
