+++ swumeal/src/main/java/com/swumeal/app/domain/menu/service/MenuService.java
@@ -1,7 +1,7 @@
-package com.swumeal.app.service;
+package com.swumeal.app.domain.menu.service;
 
-import com.swumeal.app.dao.MenuDAO;
-import com.swumeal.app.domain.dto.MenuDTO;
+import com.swumeal.app.domain.menu.domain.MenuDAO;
+import com.swumeal.app.domain.menu.dto.MenuDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,11 +13,11 @@ public class MenuService {
 
     private final MenuDAO menuDAO;
 
-    public List<MenuDTO> todaygetList(){
+    public List<MenuDTO> todaygetList() {
         return menuDAO.find();
     }
 
-    public List<MenuDTO> getListAll(){
+    public List<MenuDTO> getListAll() {
         return menuDAO.findAll();
     }
 }
