+++ swumeal/src/test/java/com/swumeal/app/service/MenuServiceTests.java
@@ -1,12 +1,11 @@
 package com.swumeal.app.service;
 
-import com.swumeal.app.domain.dto.MenuDTO;
+import com.swumeal.app.domain.menu.dto.MenuDTO;
+import com.swumeal.app.domain.menu.service.MenuService;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import java.util.List;
 
 @SpringBootTest
 @Slf4j
@@ -16,11 +15,12 @@ public class MenuServiceTests {
     private MenuService menuService;
 
     @Test
-    public void todaygetListTest(){
+    public void todaygetListTest() {
         menuService.todaygetList().stream().map(MenuDTO::toString).forEach(log::info);
     }
+
     @Test
-    public void getListAllTest(){
+    public void getListAllTest() {
         menuService.getListAll().stream().map(MenuDTO::toString).forEach(log::info);
     }
 }
